Refactoring + lint + tests
